refactor #41: 인증목록 페이지 문제 해결 (3)
- kingfisher 라이브러리 적용 (이미지 캐싱)
